+++ geoips/interfaces/algorithms.py
@@ -4,6 +4,16 @@ from geoips.interfaces.base import BaseInterface, BasePlugin
 class AlgorithmsInterface(BaseInterface):
     name = "algorithms"
     deprecated_family_attr = "alg_func_type"
+    required_args = {'single_channel': ['arrays'],
+                     'channel_combination': ['arrays'],
+                     'list_numpy_to_numpy': ['arrays'],
+                     'xarray_to_numpy': ['xobj'],
+                     'rgb': ['arrays'],
+                     'xarray_dict_to_xarray': ['xarray_dict'],
+                     'xarray_dict_dict_to_xarray': ['xarray_dict_dict'],
+                     'xarray_dict_to_xarray_dict': ['xarray_dict'],
+                     'xarray_dict_area_def_to_numpy': ['xarray_dict', 'area_def'],
+                     }
 
 
 algorithms = AlgorithmsInterface()