+++ geoips/interfaces/procflows.py
@@ -4,6 +4,7 @@ from geoips.interfaces.base import BaseInterface, BasePlugin
 class ProcflowsInterface(BaseInterface):
     name = "procflows"
     deprecated_family_attr = "procflow_type"
+    required_args = {'standard': ['fnames']}
 
 
 procflows = ProcflowsInterface()