+++ geoips/interfaces/title_formats.py
@@ -5,6 +5,7 @@ class TitleFormattersInterface(BaseInterface):
     name = "title_formats"
     entry_point_group = "title_formats"
     deprecated_family_attr = "title_type"
+    required_args = {'standard': []}
 
 
 title_formats = TitleFormattersInterface()